Improved netcode
Former-commit-id: 590265a26946347a545be09fc9861e66cb223608

+++ Assets/Scripts/Core/IPlayer.cs
@@ -4,7 +4,7 @@ namespace BoM.Core {
 	public interface IPlayer {
 		Account Account { get; }
 		ulong ClientId { get; }
-		long Latency { get; }
+		float Latency { get; }
 		Vector3 RemotePosition { get; set; }
 		Vector3 RemoteDirection { get; set; }
 	}
